+++ src/presets/index.ts
@@ -631,3 +631,4 @@ export function publicApiPreset(
     ...options,
   };
 }
+